ext-x-define: review rewrite
