+++ src/rs-core/parser/mod.rs
@@ -1,18 +1,21 @@
+mod attribute_list;
 mod audio_track_list;
 mod media_playlist;
 mod media_tag;
 mod multi_variant_playlist;
-mod utils;
+mod value_parsers;
+mod variable_substitution;
 mod variant_stream;
 
 pub(crate) use audio_track_list::AudioTrack;
 pub(crate) use media_playlist::{
-    ByteRange, InitSegmentInfo, MediaPlaylist, MediaPlaylistParsingError, MediaSegmentInfo,
-    SegmentList, SegmentTimeInfo,
+    InitSegmentInfo, MediaPlaylist, MediaPlaylistParsingError, MediaSegmentInfo, SegmentList,
+    SegmentTimeInfo,
 };
 pub(crate) use media_tag::{MediaTag, MediaTagType};
 pub(crate) use multi_variant_playlist::{
     MediaPlaylistPermanentId, MediaPlaylistUpdateError, MultivariantPlaylist,
     MultivariantPlaylistParsingError,
 };
+pub(crate) use value_parsers::ByteRange;
 pub(crate) use variant_stream::{VariantStream, VideoResolution};
